% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_args.R
\name{covariate_args}
\alias{covariate_args}
\title{Assigns common covariate arguments to argparse object}
\usage{
covariate_args(parser, ..., output_root = "<covariate root>")
}
\arguments{
\item{parser}{argparse::ArgumentParser instance or equivalent}
}
\description{
Assigns common covariate arguments to argparse object
}
