% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{.get.shp.metadata}
\alias{.get.shp.metadata}
\title{Return extra metadata associated with .shp file}
\usage{
.get.shp.metadata(path)
}
\description{
A "shapefile" is actually a collection of files. At minimum .shp, .shpx, and .dbf
}
\details{
In addition, there are at least 13 optional files that may be included.

https://en.wikipedia.org/wiki/Shapefile#Overview
}
