% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{.replace.methods.with.spy}
\alias{.replace.methods.with.spy}
\title{Replace methods in global namespace with the spying methods defined below}
\usage{
.replace.methods.with.spy(methods)
}
\description{
Replace methods in global namespace with the spying methods defined below
}
