% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_on_missing_locations.R
\name{error_on_missing_locations}
\alias{error_on_missing_locations}
\title{Error on missing locations in hierarchy}
\usage{
error_on_missing_locations(
  data,
  ...,
  hierarchy,
  hierarchy.value = "location_id"
)
}
\arguments{
\item{data}{vector-like with values corresponding to the location hierarchy}

\item{...}{captures other arguments. Do not provide any or you will get an error.}

\item{hierarchy}{IHME location hierarchy from e.g., \code{get_location_metadata}}

\item{hierarchy.value}{the value in the hierarchy to compare to. Default: "location_id"}
}
\description{
Compares data to location hierarchy to ensure that data has values present for each hierarchy entry.
Issues a \code{stop} if any values are missing.
}
