% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_output_date_index.R
\name{get_latest_output_date_index}
\alias{get_latest_output_date_index}
\title{get the latest index for given an output dir and a date}
\usage{
get_latest_output_date_index(dir, date)
}
\arguments{
\item{dir}{path to directory with versioned dirs}

\item{date}{character in be YYYY_MM_DD format}
}
\value{
largest version in directory tree or 0 if there are no version OR
the directory tree does not exist
}
\description{
directories are assumed to be named in YYYY_MM_DD.VV format with sane
year/month/date/version values.
}
