% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_dir.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get output directory for results to save in}
\usage{
get_output_dir(root, date)
}
\arguments{
\item{root}{path to root of output results}

\item{date}{character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}
}
\description{
Returns an appropriate path to save results in, creating it if necessary.
}
