% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_dir.R
\name{get_script_dir}
\alias{get_script_dir}
\title{Returns the directory the currently running script is in or NULL if all fails}
\usage{
get_script_dir()
}
\description{
Returns the directory the currently running script is in or NULL if all fails
}
