% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_of_children.R
\name{parent_of_child}
\alias{parent_of_child}
\title{parent_of_child}
\usage{
parent_of_child(child_location_id, hierarchy, parent_level)
}
\arguments{
\item{child_location_id}{[int] Location ID to pull parent ID for}

\item{hierarchy}{[data.table] Hierarchy. Must have columns location_id, path_to_top_parent, and level.}

\item{parent_level}{[int] Single level of the hierarchy - find all parent location_ids of child location_ids at this level}
}
\value{

}
\description{
Given a location ID, returns its parent ID at a given level.
}
