% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_of_children.R
\name{parents_of_children}
\alias{parents_of_children}
\title{parents_of_children}
\usage{
parents_of_children(child_loc_ids, hierarchy, parent_level)
}
\arguments{
\item{child_loc_ids}{[int] Vector of location IDs to pull parent ID for}

\item{hierarchy}{[data.table] Hierarchy. Must have columns location_id, path_to_top_parent, and level.}

\item{parent_level}{[int] Single level of the hierarchy - find all parent location_ids of child location_ids at this level}
}
\description{
Vectorized implementation of "parents_of_child"
}
