% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_array_job.R
\name{pdf_array_job}
\alias{pdf_array_job}
\title{Submit an array job to plot PDFs, then bind back together.}
\usage{
pdf_array_job(
  jobs_file,
  r_script,
  final_out_dir,
  write_file_name = "all_jobs.pdf",
  rshell = "/ihme/singularity-images/rstudio/shells/execRscript.sh -s",
  fthread = 5,
  fmem = "10G",
  h_rt = "02:00:00",
  queue = "all.q",
  project = "proj_covid",
  job_name = "pdf_array",
  output = NULL,
  errors = NULL,
  remove_temp_files = TRUE,
  args = NULL
)
}
\arguments{
\item{jobs_file}{.csv that has one row for each job you want to run, with arguments for plots as columns.}

\item{r_script}{Full path to the R script that has your plotting code.}

\item{final_out_dir}{Path to directory to save final file. Needs to be to a path on the cluster.}

\item{write_file_name}{Final file name to write. Defaults to "all_jobs.pdf".}

\item{rshell}{R-shell to use. Defaults to SciComp shell.}

\item{fthread}{fthreads for cluster jobs. Defaults to 5.}

\item{fmem}{fmem for cluster jobs. Defaults to 10G.}

\item{h_rt}{h_rt for cluster jobs. Defaults to 2 hours.}

\item{queue}{Queue for cluster jobs. Options are 'all.q', 'i.q', 'long.q', or 'd.q'. Defaults to 'all.q'.}

\item{project}{Cluster project. Defaults to 'proj_covid'.}

\item{job_name}{Cluster job name. Defaults to 'pdf_array'.}

\item{output}{Path to write cluster output messages to. Defaults to "/ihme/temp/slurmoutput/USERNAME/output"}

\item{errors}{Path to write cluster error messages to. Defaults to "/ihme/temp/slurmoutput/USERNAME/errors"}

\item{remove_temp_files}{Remove temporary files after jobs have completed? Defaults to TRUE.}

\item{args}{Additional arguments to pass to array job launch script.}
}
\description{
Submit an array job to plot PDFs, then bind back together.
}
