% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_debug.R
\name{print_debug}
\alias{print_debug}
\title{Convenience function for printing information}
\usage{
print_debug(...)
}
\arguments{
\item{...}{variables to debug. Provide by name.}
}
\description{
Provide any number of arguments which are the variables you want printed
Note that this function must be called directly and not wrapped in e.g., capture.output
}
\examples{
\dontrun{
foo <- 7
bar <- 9
print_debug(foo, bar)
}
}
