% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect_defaults.R
\name{reflect_defaults}
\alias{reflect_defaults}
\title{Reflect default configuration values for ihme.covid based off of filesystem}
\usage{
reflect_defaults()
}
\description{
Reflect default configuration values for ihme.covid based off of filesystem
}
