% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_hierarchy.R
\name{sort_hierarchy}
\alias{sort_hierarchy}
\title{Sort data.table hierarchy for easier use of downstream outputs}
\usage{
sort_hierarchy(dt, prepend = NULL)
}
\arguments{
\item{dt}{data.table instance. MUST have the following columns: location_id, path_to_top_parent}

\item{prepend}{numeric vector of location_id values to prepend to the document. If provided, these values
will be moved to the beginning of the result in the order provided and all other values will be sorted.}
}
\description{
Sorts the hierarchy by location hierarchy so that children follow parents and everything is alphabetical.
In addition, arbitrary locations can be moved to the front for easy lookup in e.g., output PDFs.
}
