% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_covid_shapes.R
\name{load_covid_shapes}
\alias{load_covid_shapes}
\title{Load COVID-19 shapes from standard model-inputs}
\usage{
load_covid_shapes(location_set_id, location_set_version_id, shp_path = NULL)
}
\arguments{
\item{location_set_id}{integer representing location_set_id of the hierarchy}

\item{location_set_version_id}{integer representing location_set_version_id of the hierarchy}

\item{shp_path}{path to shapefile to load. Used for testing. Default: NULL, which will use the shapefile of "best"}
}
\description{
Load COVID-19 shapes and standardize column names to harmonize with standard location hierarchy names.
}
